pyproject.toml
[tool.poetry]
name = "gainy-compute"
version = "0.5.27"
description = "All stuff related to data computations and ML models in Python"
authors = [
    "Vasilii Khomutov <vasilii@gainy.app>",
    "Mikhail Astashkevich <mikhail@gainy.app>",
]

packages = [
    { include = "gainy"},
]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
pandas = "^1.3.4"
scipy = "^1.7.0,<1.9"
boto3 = "^1.20.24"
psycopg2-binary = "^2.9.3"
nltk = "^3.6.7"
numpy = "^1.22.1"
backoff = "^2.2.1"
python-json-logger = "^2.0.4"
stripe = "^4.1.0"
"github3.py" = "^3.2.0"
scikit-learn = "^1.1.3"
requests = "^2.28.1"
plaid-python = "^11.0.0"
amplitude-analytics = "^1.1.1"
pytz = "^2022.7.1"
sendgrid = "^6.9.7"
appsflyer = "^0.1.0"
slack-sdk = "^3.21.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# ML
gainy_industry_assignment = "gainy.industries.runner:cli"
gainy_recommendation = "gainy.recommendation.job:cli"
gainy_optimize_collections = "gainy.optimization.jobs.optimize_collections:cli"

# Trading
gainy_update_account_balances = "gainy.trading.jobs.update_account_balances:cli"
gainy_rebalance_portfolios = "gainy.trading.drivewealth.jobs.rebalance_portfolios:cli"
gainy_fetch_drivewealth_instruments = "gainy.trading.drivewealth.jobs.fetch_instruments:cli"
gainy_fetch_drivewealth_countries = "gainy.trading.drivewealth.jobs.fetch_countries:cli"

# Billing
gainy_create_invoices = "gainy.billing.jobs.create_invoices:cli"
gainy_charge_invoices = "gainy.billing.jobs.charge_invoices:cli"

# Misc
gainy_github_update_file = "gainy.github.update_file_job:cli"
gainy_sync_profiles_analytics_attributes = "gainy.analytics.jobs.sync_profiles_attributes:cli"
gainy_rehandle_failed_events = "gainy.queue_processing.jobs.rehandle_failed_events:cli"
